refactor: footer and about components
