+++ src/App.css
@@ -1,38 +1,31 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+.page {
+  margin: 0;
+  padding: 0;
+  background-color: #fff;
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  margin: auto;
+  min-height: 100vh;
+  width: 100%;
+  max-width: 1440px;
 }
 
-.App-link {
-  color: #61dafb;
+.page__about-author {
+  margin: 80px 0;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
+@media screen and (max-width: 768px) {
+  .page__about-author {
+    margin: 61px 0;
   }
-  to {
-    transform: rotate(360deg);
+}
+
+@media screen and (max-width: 550px) {
+  .page {
+    align-items: center;
+  }
+  .page__about-author {
+    margin: 32px 0;
   }
 }
