+++ src/App.js
@@ -1,8 +1,8 @@
 import "./App.css";
 import Header from "./components/Header/Header";
 import Main from "./components/Main/Main";
-import Footer from "./components/Footer";
-import About from "./components/About";
+import Footer from "./components/Footer/Footer";
+import About from "./components/About/About";
 import newsApi from "./utils/newsApi";
 import { useEffect, useState } from "react";
 import SignInModalForm from "./components/SignInModalForm/SignInModalForm";
