--- src/blocks/about-author/__description/about-author__description.css
@@ -1,10 +1,0 @@
-.about-author__description {
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  row-gap: 24px;
-  font-family: "Roboto";
-  font-size: 18px;
-  line-height: 24px;
-}
