--- src/blocks/about-author/__paragraph/about-author__paragraph.css
@@ -1,4 +1,0 @@
-.about-author__paragraph {
-  margin: 0;
-  padding: 0;
-}
