--- src/blocks/about-author/__photo-container/about-author__photo-container.css
@@ -1,19 +1,0 @@
-.about-author__photo-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  grid-area: photo;
-  width: 100%;
-}
-
-@media screen and (max-width: 768px) {
-  .about-author__photo-container {
-    width: 232px;
-  }
-}
-
-@media screen and (max-width: 550px) {
-  .about-author__photo-container {
-    width: 272px;
-  }
-}
