--- src/blocks/about-author/__photo-mask/about-author__photo-mask.css
@@ -1,12 +1,0 @@
-.about-author__photo-mask {
-  grid-area: photo;
-  overflow: hidden;
-  border-radius: 50%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  position: relative;
-  padding-bottom: 100%;
-  justify-self: center;
-}
