--- src/blocks/about-author/__photo/about-author__photo.css
@@ -1,8 +1,0 @@
-.about-author__photo {
-  object-fit: cover;
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  left: 0;
-  top: 0;
-}
