--- src/blocks/about-author/__title/about-author__title.css
@@ -1,8 +1,0 @@
-.about-author__title {
-  margin: 0;
-  padding: 0;
-  font-family: "Roboto Slab";
-  font-size: 40px;
-  font-weight: 400;
-  line-height: 46px;
-}
