--- src/blocks/about-author/about-author.css
@@ -1,38 +1,0 @@
-.about-author {
-  margin: 0;
-  padding: 0 64px;
-  display: grid;
-  grid-template-columns: minmax(232px, 423px) auto;
-  grid-template-rows: auto;
-  grid-template-areas: "photo info";
-  column-gap: 56px;
-  max-width: calc(100% - 208px);
-  width: 100%;
-  align-items: start;
-}
-
-@media screen and (max-width: 1070px) {
-  .about-author {
-    max-width: calc(100% - 80px);
-    padding: 0;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .about-author {
-    grid-template-columns: min-content 1fr;
-    justify-content: center;
-    column-gap: 32px;
-  }
-}
-
-@media screen and (max-width: 550px) {
-  .about-author {
-    justify-items: center;
-    grid-template-columns: auto;
-    grid-template-areas:
-      "photo"
-      "info";
-    column-gap: 32px;
-  }
-}
