--- src/blocks/footer/__container/footer__container.css
@@ -1,23 +1,0 @@
-.footer__container {
-  width: 100%;
-  grid-area: container;
-  display: grid;
-  grid-template-columns: auto auto;
-  grid-template-rows: auto;
-  grid-template-areas: "copyright options";
-  align-items: center;
-  justify-content: space-between;
-}
-
-@media screen and (max-width: 690px) {
-  .footer__container {
-    margin-top: 20px;
-    margin-bottom: 20px;
-    row-gap: 38px;
-    grid-template-columns: 1fr;
-    grid-template-rows: auto;
-    grid-template-areas:
-      "options"
-      "copyright";
-  }
-}
