--- src/blocks/footer/__copyright/footer__copyright.css
@@ -1,17 +1,0 @@
-.footer__copyright {
-  margin: 0;
-  padding: 0;
-  color: #b6bcbf;
-  font-family: "Roboto";
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 22px;
-  grid-area: copyright;
-}
-
-@media screen and (max-width: 690px) {
-  .footer__copyright {
-    display: flex;
-    justify-content: center;
-  }
-}
