--- src/blocks/footer/__icon/footer__icon.css
@@ -1,5 +1,0 @@
-.footer__icon {
-  margin: 0;
-  padding: 0;
-  margin: 2px;
-}
