--- src/blocks/footer/__link-icon/footer__link-icon.css
@@ -1,9 +1,0 @@
-.footer__link-icon {
-  margin: 0;
-  padding: 0;
-  width: 24px;
-  height: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
