--- src/blocks/footer/__link-icons/footer__link-icons.css
@@ -1,13 +1,0 @@
-.footer__link-icons {
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  column-gap: 24px;
-}
-
-@media screen and (max-width: 690px) {
-  .footer__link-icons {
-    justify-content: end;
-  }
-}
