--- src/blocks/footer/__link/footer__link.css
@@ -1,10 +1,0 @@
-.footer__link {
-  margin: 0;
-  padding: 0;
-  font-family: "Roboto";
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 24px;
-  text-decoration: none;
-  color: #1a1b22;
-}
