--- src/blocks/footer/__links/footer__links.css
@@ -1,15 +1,0 @@
-.footer__links {
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  column-gap: 43px;
-}
-
-@media screen and (max-width: 690px) {
-  .footer__links {
-    flex-direction: column;
-    row-gap: 38px;
-    justify-content: space-between;
-  }
-}
