--- src/blocks/footer/footer.css
@@ -1,27 +1,0 @@
-.footer {
-  width: calc(100% - 32px);
-  padding: 0 16px;
-  max-width: 1200px;
-  display: grid;
-  align-items: center;
-  min-height: 80px;
-  grid-template-columns: auto;
-  grid-template-rows: auto;
-  grid-template-areas: "container";
-}
-
-@media screen and (max-width: 768px) {
-  .footer {
-    min-height: 64px;
-    width: calc(100% - 80px);
-    padding: 0 40px;
-    max-width: 1200px;
-  }
-}
-
-@media screen and (max-width: 690px) {
-  .footer {
-    width: calc(100% - 32px);
-    padding: 0 16px;
-  }
-}
