--- src/blocks/page/__about-author/page__about-author.css
@@ -1,15 +1,0 @@
-.page__about-author {
-  margin: 80px 0;
-}
-
-@media screen and (max-width: 768px) {
-  .page__about-author {
-    margin: 61px 0;
-  }
-}
-
-@media screen and (max-width: 550px) {
-  .page__about-author {
-    margin: 32px 0;
-  }
-}
