--- src/blocks/page/page.css
@@ -1,18 +1,0 @@
-.page {
-  margin: 0;
-  padding: 0;
-  background-color: #fff;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: auto;
-  min-height: 100vh;
-  width: 100%;
-  max-width: 1440px;
-}
-
-@media screen and (max-width: 550px) {
-  .page {
-    align-items: center;
-  }
-}
