+++ src/components/Footer/Footer.css
@@ -1,0 +1,128 @@
+.footer {
+  width: calc(100% - 32px);
+  padding: 0 16px;
+  max-width: 1200px;
+  display: grid;
+  align-items: center;
+  min-height: 80px;
+  grid-template-columns: auto;
+  grid-template-rows: auto;
+  grid-template-areas: "container";
+}
+
+.footer__container {
+  width: 100%;
+  grid-area: container;
+  display: grid;
+  grid-template-columns: auto auto;
+  grid-template-rows: auto;
+  grid-template-areas: "copyright options";
+  align-items: center;
+  justify-content: space-between;
+}
+
+.footer__copyright {
+  margin: 0;
+  padding: 0;
+  color: #b6bcbf;
+  font-family: "Roboto";
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 22px;
+  grid-area: copyright;
+}
+
+.footer__icon {
+  margin: 0;
+  padding: 0;
+  margin: 2px;
+}
+
+.footer__link {
+  margin: 0;
+  padding: 0;
+  font-family: "Roboto";
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 24px;
+  text-decoration: none;
+  color: #1a1b22;
+}
+
+.footer__link-icon {
+  margin: 0;
+  padding: 0;
+  width: 24px;
+  height: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.footer__link-icons {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: row;
+  column-gap: 24px;
+}
+
+.footer__links {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: row;
+  column-gap: 43px;
+}
+
+.footer__options {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: row;
+  column-gap: 40px;
+  grid-area: options;
+}
+
+@media screen and (max-width: 768px) {
+  .footer {
+    min-height: 64px;
+    width: calc(100% - 80px);
+    padding: 0 40px;
+    max-width: 1200px;
+  }
+}
+
+@media screen and (max-width: 690px) {
+  .footer__options {
+    display: grid;
+    grid-template-columns: auto auto;
+    width: 100%;
+  }
+  .footer__links {
+    flex-direction: column;
+    row-gap: 38px;
+    justify-content: space-between;
+  }
+  .footer__link-icons {
+    justify-content: end;
+  }
+  .footer__copyright {
+    display: flex;
+    justify-content: center;
+  }
+  .footer__container {
+    margin-top: 20px;
+    margin-bottom: 20px;
+    row-gap: 38px;
+    grid-template-columns: 1fr;
+    grid-template-rows: auto;
+    grid-template-areas:
+      "options"
+      "copyright";
+  }
+  .footer {
+    width: calc(100% - 32px);
+    padding: 0 16px;
+  }
+}
