+++ src/index.css
@@ -2,31 +2,12 @@
 @import url("./fonts/roboto/roboto.css");
 @import url("./fonts/inter/inter.css");
 @import url("./fonts/source_sans_pro/source_sans_pro.css");
-@import url("./blocks/page/page.css");
 @import url("./blocks/button/button.css");
 @import url("./blocks/button/__content/button__content.css");
 @import url("./blocks/button/__icon/button__icon.css");
 @import url("./blocks/button/__icon-container/button__icon-container.css");
 @import url("./blocks/button/__text/button__text.css");
 @import url("./blocks/link-button/link-button.css");
-@import url("./blocks/about-author/about-author.css");
-@import url("./blocks/page/__about-author/page__about-author.css");
-@import url("./blocks/about-author/__photo-container/about-author__photo-container.css");
-@import url("./blocks/about-author/__photo-mask/about-author__photo-mask.css");
-@import url("./blocks/about-author/__photo/about-author__photo.css");
-@import url("./blocks/about-author/__info/about-author__info.css");
-@import url("./blocks/about-author/__title/about-author__title.css");
-@import url("./blocks/about-author/__description/about-author__description.css");
-@import url("./blocks/about-author/__paragraph/about-author__paragraph.css");
-@import url("./blocks/footer/footer.css");
-@import url("./blocks/footer/__container/footer__container.css");
-@import url("./blocks/footer/__copyright/footer__copyright.css");
-@import url("./blocks/footer/__options/footer__options.css");
-@import url("./blocks/footer/__links/footer__links.css");
-@import url("./blocks/footer/__link/footer__link.css");
-@import url("./blocks/footer/__link-icons/footer__link-icons.css");
-@import url("./blocks/footer/__link-icon/footer__link-icon.css");
-@import url("./blocks/footer/__icon/footer__icon.css");
 @import url("./blocks/modal/modal.css");
 @import url("./blocks/modal/_state/modal_state_open.css");
 @import url("./blocks/modal/_state/modal_state_close.css");
